List posyandu && add migration 'ibu hamil'

+++ database/migrations/2022_07_25_164406_create_pemeriksaan_balita_table.php
@@ -18,10 +18,10 @@ class CreatePemeriksaanBalitaTable extends Migration
             $table->bigInteger('id_balita')->length(20)->unsigned();
             $table->bigInteger('id_posyandu')->length(20)->unsigned();
             $table->bigInteger('id_user_petugas')->length(20)->unsigned();
-            $table->integer('berat_badan')->length(20);
-            $table->integer('tinggi_badan')->length(20);
-            $table->integer('lingkar_lengan')->length(20);
-            $table->integer('lingkar_kepala')->length(20);
+            $table->float('berat_badan')->length(20);
+            $table->float('tinggi_badan')->length(20);
+            $table->float('lingkar_lengan')->length(20);
+            $table->float('lingkar_kepala')->length(20);
             $table->tinyInteger('is_deleted')->default(0);
             $table->timestamps();
         });
